Move FailingAppTokenValidator to FizzServerHandshake.cpp
