+++ quic/server/handshake/AppToken.cpp
@@ -8,21 +8,6 @@
 
 #include <quic/server/handshake/AppToken.h>
 
-#include <quic/QuicConstants.h>
-#include <quic/fizz/handshake/FizzTransportParameters.h>
-
-#include <fizz/record/Types.h>
-#include <folly/IPAddress.h>
-#include <folly/Optional.h>
-#include <folly/io/Cursor.h>
-#include <folly/io/IOBuf.h>
-
-#include <glog/logging.h>
-
-#include <cstdint>
-#include <memory>
-#include <vector>
-
 namespace quic {
 
 TicketTransportParameters createTicketTransportParameters(
@@ -55,40 +40,6 @@ TicketTransportParameters createTicketTransportParameters(
   params.parameters.push_back(encodeIntegerParameter(
       TransportParameterId::initial_max_streams_uni, initialMaxStreamsUni));
   return params;
-}
-
-std::unique_ptr<folly::IOBuf> encodeAppToken(const AppToken& appToken) {
-  auto buf = folly::IOBuf::create(20);
-  folly::io::Appender appender(buf.get(), 20);
-  auto ext = encodeExtension(appToken.transportParams, QuicVersion::MVFST);
-  fizz::detail::write(ext, appender);
-  fizz::detail::writeVector<uint8_t>(appToken.sourceAddresses, appender);
-  fizz::detail::write(appToken.version, appender);
-  fizz::detail::writeBuf<uint16_t>(appToken.appParams, appender);
-  return buf;
-}
-
-folly::Optional<AppToken> decodeAppToken(const folly::IOBuf& buf) {
-  AppToken appToken;
-  folly::io::Cursor cursor(&buf);
-  std::vector<fizz::Extension> extensions;
-  fizz::Extension ext;
-  try {
-    fizz::detail::read(ext, cursor);
-    extensions.push_back(std::move(ext));
-    // TODO plumb version
-    appToken.transportParams =
-        *fizz::getTicketExtension(extensions, QuicVersion::MVFST);
-    fizz::detail::readVector<uint8_t>(appToken.sourceAddresses, cursor);
-    if (cursor.isAtEnd()) {
-      return appToken;
-    }
-    fizz::detail::read(appToken.version, cursor);
-    fizz::detail::readBuf<uint16_t>(appToken.appParams, cursor);
-  } catch (const std::exception&) {
-    return folly::none;
-  }
-  return appToken;
 }
 
 } // namespace quic
