+++ quic/server/handshake/test/AppTokenTest.cpp
@@ -6,7 +6,7 @@
  *
  */
 
-#include <quic/server/handshake/AppToken.h>
+#include <quic/fizz/server/handshake/AppToken.h>
 
 #include <quic/QuicConstants.h>
 #include <quic/server/state/ServerStateMachine.h>
